added iniphile.spec.in, rpm target
preliminary code (not production-quality):

* iniphile.spec may become stale WRT VERSIONSTRING
  (mk/common-values.mk)
* iniphile.spec.in is only tested with openSUSE

the specfile is targeted at the SUSE Build Service
so i should be able to extend support to at least RHEL
through that.

also, the openSUSE packaging guidelines tell me i should
move the header files and the static library out to
a separate iniphile-devel package, we'll see about that.